Triangles: Implement instanced rendering
Note: There are multiple new sections of code that have been left commented out. These are just the remnants of another version of instanced rendering that had some issues on my iGPU drivers. Need to see whether it 1. works on more powerful discrete PCs and 2. if it does, whether the performance difference between the two implementations is worth it.