XG/SSQ: hold model/mesh matrices in the xgDagMesh node
This goes against how the original engine (on PS2 at least) handled holding model matrices, but should remove the recalculations of model matrices when drawing.

Note: this is due to how the original game would draw both the opaque AND transparent meshes of a model all at the same time. With that implementation, there would've been no recalculations of mesh matrices - at the cost of improper alpha ordering (ex: ST07).