Viewer: Alpha blending change
Disables writes to the depth buffer by transparent objects. Most objects are placed in order, and with draw calls separate between all opaques and all transparents, this should emulate how the PSP version handled transparency.

That also comes at the cost of having some transparent objects drawn out of order... but it's accurate to the original.