SSQ: use combined mesh rendering where necessary
To emulate some effects properly as shown in the original game, some stages have to have both opaque and transparent meshes rendered in one go. This is only necessary for models that ignore depth testing.

Examples: the vortex behind Panpus in ST01 and the swirl effect before the charge>battle transition ALSO in ST01.